Update README.md
Equation of a Line

Write a program to calculate the slope between two points (x1, y1) and (x2, y2). The points should be entered as four floating-point values. Output should be rounded to 2 decimal places.

The formula to calculate the slope, m, between two points is

First, output the calculated slope value. Then, use this value to output the equation of the line in the form:

Calculate the value of c using one of the pairs of points entered as input. A sample input and output is shown below:

Let's find the equation of a line using 2 points.
Enter x1: 9
Enter y1: 4
Enter x2: -4
Enter y2: 7

The slope of the line is: -0.23
The equation of the line is: y = -0.23x + 6.08